separated logic for aggregated bundles type
